requirements.txt
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
PyYAML==5.4.1
scikit-learn==0.24.2
tensorboard==2.6.0
torch==1.10.0+cu113
tqdm==4.62.3
